requirements.txt
# VLM & Vision
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0

# CSP Solving
python-constraint>=1.4.0

# Data & ML
numpy>=1.24.0
pandas>=2.0.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Development
pytest>=7.3.0
black>=23.0.0
ruff>=0.0.270

kaggle